Separate components, Fix base on feedback
