+++ src/interfaces/hardware.d.ts
@@ -355,3 +355,65 @@ export interface IAssetsWaiting {
   location_id: number;
   assigned_status: number;
 }
+
+export interface ISearchFormProps {
+  searchFormProps: any;
+  locationSelectProps: any;
+  handleChangePickerByMonth: (val: any, formatString: any) => void;
+  handleChangeLocation: (value: number) => void;
+  searchValuesLocation: number;
+  searchValuesByDateFrom: string | undefined;
+  searchValuesByDateTo: string | undefined;
+  rtd_location_id: string | null;
+  dateFromParam: string | null;
+  dateToParam: string | null;
+}
+
+export interface ITableProps {
+  columns: any[];
+  selectedColumns: string[];
+  tableProps: any;
+  onShow: (record: IHardwareResponse) => void;
+  onEdit: (record: IHardwareResponse) => void;
+  onDeleteSuccess: () => void;
+  pageTotal: number | false | undefined;
+  t: any;
+  resourceName?: string;
+}
+
+export interface IToolBarProps {
+  columns: ColumnItem[];
+  selectedColumns: string[];
+  onToggleColumn: (col: ColumnItem) => void;
+  onRefresh: () => void;
+  onOpenSearch: () => void;
+  t: any;
+}
+
+export interface IModalPropsProps {
+  t: any;
+
+  isModalVisible: boolean;
+  setIsModalVisible: (val: boolean) => void;
+
+  isEditModalVisible: boolean;
+  setIsEditModalVisible: (val: boolean) => void;
+
+  isShowModalVisible: boolean;
+  setIsShowModalVisible: (val: boolean) => void;
+
+  isSearchModalVisible: boolean;
+  setIsSearchModalVisible: (val: boolean) => void;
+
+  isCheckoutModalVisible: boolean;
+  setIsCheckoutModalVisible: (val: boolean) => void;
+
+  isCheckinModalVisible: boolean;
+  setIsCheckinModalVisible: (val: boolean) => void;
+
+  detail: IHardwareResponse | undefined;
+  detailCheckin: IHardwareResponseCheckin | undefined;
+  detailCheckout: IHardwareResponseCheckout | undefined;
+
+  searchFormProps: any;
+}
