+++ src/pages/hardware/list-maintenance.tsx
@@ -6,49 +6,16 @@ import {
   useNavigation,
   usePermissions,
 } from "@pankod/refine-core";
-import {
-  List,
-  Table,
-  TextField,
-  useTable,
-  getDefaultSortOrder,
-  DateField,
-  Space,
-  CloneButton,
-  EditButton,
-  DeleteButton,
-  TagField,
-  CreateButton,
-  Button,
-  ShowButton,
-  Tooltip,
-  Checkbox,
-  useSelect,
-  Form,
-  Select,
-} from "@pankod/refine-antd";
-import { Image, Spin, DatePicker } from "antd";
+import { List, useTable, CreateButton, useSelect } from "@pankod/refine-antd";
+import { Spin } from "antd";
 import "styles/antd.less";
-import { useEffect, useMemo, useRef, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import { useSearchParams } from "react-router-dom";
 import moment from "moment";
 
-import { IHardware } from "interfaces";
 import { TableAction } from "components/elements/tables/TableAction";
-import { MModal } from "components/Modal/MModal";
-import { HardwareCreate } from "./create";
-import { HardwareEdit } from "./edit";
-import { HardwareShow } from "./show";
-import { HardwareCheckout } from "./checkout";
-import { HardwareCheckin } from "./checkin";
-import { HardwareSearch } from "./search";
-import { TotalDetail } from "components/elements/TotalDetail";
 
-import {
-  MenuOutlined,
-  FileSearchOutlined,
-  SyncOutlined,
-} from "@ant-design/icons";
+import { TotalDetail } from "components/elements/TotalDetail";
 
 import {
   IHardwareFilterVariables,
@@ -64,21 +31,15 @@ import {
   HARDWARE_TOTAL_DETAIL_API,
 } from "api/baseApi";
 import { ICompany } from "interfaces/company";
-import {
-  dateFormat,
-  STATUS_LABELS,
-  STATUS_MAINTENANCE,
-} from "constants/assets";
-import {
-  filterAssignedStatus,
-  getAssetAssignedStatusDecription,
-  getAssetStatusDecription,
-  getBGAssetAssignedStatusDecription,
-  getBGAssetStatusDecription,
-} from "untils/assets";
 import { ICategory } from "interfaces/categories";
 import { IStatusLabel } from "interfaces/statusLabel";
 import { EPermissions } from "constants/permissions";
+import { useHardwareColumns } from "./table-column";
+import { SearchFilterForm } from "./search-filter-form";
+import { ToolbarActions } from "./tool-bar";
+import { HardwareTable } from "./table";
+import { ModalsWrapper } from "./modal";
+import { convertHardwareToEditData } from "ultils/ConvertHardwareData";
 
 const defaultCheckedList = [
   "id",
@@ -102,38 +63,27 @@ export const HardwareListMaintenance: React.FC<
   const [detailCheckout, setDetailCheckout] =
     useState<IHardwareResponseCheckout>();
   const [isCheckoutModalVisible, setIsCheckoutModalVisible] = useState(false);
-  const [isLoadingArr] = useState<boolean[]>([]);
-  const [isCloneModalVisible, setIsCloneModalVisible] = useState(false);
   const [isShowModalVisible, setIsShowModalVisible] = useState(false);
   const [isCheckinModalVisible, setIsCheckinModalVisible] = useState(false);
   const [detailCheckin, setDetailCheckin] =
     useState<IHardwareResponseCheckin>();
-  const [detailClone, setDetailClone] = useState<IHardwareResponse>();
   const [collumnSelected, setColumnSelected] = useState<string[]>(
     localStorage.getItem("item_selected_maintenance") !== null
       ? JSON.parse(localStorage.getItem("item_selected_maintenance") as any)
       : defaultCheckedList
   );
-  const [isActive, setIsActive] = useState(false);
-  const onClickDropDown = () => setIsActive(!isActive);
-  const menuRef = useRef(null);
-  const [listening, setListening] = useState(false);
   const [isSearchModalVisible, setIsSearchModalVisible] = useState(false);
   const [searchParams, setSearchParams] = useSearchParams();
   const rtd_location_id = searchParams.get("rtd_location_id");
   const dateFromParam = searchParams.get("dateFrom");
   const dateToParam = searchParams.get("dateTo");
   const searchParam = searchParams.get("search");
-  const [isAdmin, setIsAdmin] = useState(false);
   const { data: permissionsData } = usePermissions();
 
-  useEffect(() => {
-    if (permissionsData.admin === EPermissions.ADMIN) {
-      setIsAdmin(true);
-    } else {
-      setIsAdmin(false);
-    }
-  }, [permissionsData]);
+  const isAdmin = useMemo(
+    () => permissionsData?.admin === EPermissions.ADMIN,
+    [permissionsData]
+  );
 
   const { tableProps, sorter, searchFormProps, tableQueryResult, filters } =
     useTable<IHardwareResponse, HttpError, IHardwareFilterVariables>({
@@ -212,113 +162,7 @@ export const HardwareListMaintenance: React.FC<
     });
 
   const edit = (data: IHardwareResponse) => {
-    const dataConvert: IHardwareResponse = {
-      id: data.id,
-      name: data.name,
-      asset_tag: data.asset_tag,
-      serial: data.serial ?? "",
-      model: {
-        id: data?.model?.id,
-        name: data?.model?.name,
-      },
-      model_number: data?.order_number,
-      status_label: {
-        id: data?.status_label.id,
-        name: data?.status_label.name,
-        status_type: data?.status_label.status_type,
-        status_meta: data?.status_label.status_meta,
-      },
-      category: {
-        id: data?.category?.id,
-        name: data?.category?.name,
-      },
-      supplier: {
-        id: data?.supplier?.id,
-        name: data?.supplier?.name,
-      },
-      notes: data.notes,
-      order_number: data.order_number ?? "",
-      location: {
-        id: data?.location?.id,
-        name: data?.location?.name,
-      },
-      rtd_location: {
-        id: data?.rtd_location?.id,
-        name: data?.rtd_location?.name,
-      },
-      image: data?.image,
-      warranty_months: data?.warranty_months,
-      purchase_cost: data?.purchase_cost,
-      purchase_date: {
-        date: data?.purchase_date !== null ? data?.purchase_date.date : "",
-        formatted:
-          data?.purchase_date !== null ? data?.purchase_date.formatted : "",
-      },
-      assigned_to: data?.assigned_to,
-      last_audit_date: data?.last_audit_date,
-
-      requestable: data?.requestable,
-      physical: data?.physical,
-
-      note: "",
-      expected_checkin: {
-        date: "",
-        formatted: "",
-      },
-      last_checkout: {
-        date: "",
-        formatted: "",
-      },
-      assigned_location: {
-        id: 0,
-        name: "",
-      },
-      assigned_user: 0,
-      assigned_asset: "",
-      checkout_to_type: {
-        assigned_user: 0,
-        assigned_asset: "",
-        assigned_location: {
-          id: 0,
-          name: "",
-        },
-      },
-      user_can_checkout: false,
-      assigned_status: 0,
-      checkin_at: {
-        date: "",
-        formatted: "",
-      },
-      created_at: {
-        datetime: "",
-        formatted: "",
-      },
-      updated_at: {
-        datetime: "",
-        formatted: "",
-      },
-      manufacturer: {
-        id: 0,
-        name: "",
-      },
-      checkin_counter: 0,
-      checkout_counter: 0,
-      requests_counter: 0,
-      warranty_expires: {
-        date: "",
-        formatted: "",
-      },
-      user_can_checkin: false,
-      withdraw_from: data?.withdraw_from,
-      maintenance_date: {
-        date: data?.maintenance_date != null ? data?.maintenance_date.date : "",
-        formatted:
-          data?.maintenance_date != null
-            ? data?.maintenance_date.formatted
-            : "",
-      },
-      maintenance_cycle: data?.maintenance_cycle,
-    };
+    const dataConvert = convertHardwareToEditData(data);
     setDetail(dataConvert);
     setIsEditModalVisible(true);
   };
@@ -358,278 +202,13 @@ export const HardwareListMaintenance: React.FC<
 
   const { list } = useNavigation();
 
-  const collumns = useMemo(
-    () => [
-      {
-        key: "id",
-        title: "ID",
-        render: (value: number) => <TextField value={value} />,
-        defaultSortOrder: getDefaultSortOrder("id", sorter),
-      },
-      {
-        key: "name",
-        title: t("hardware.label.field.assetName"),
-        render: (value: string) => <TextField value={value ? value : 0} />,
-        defaultSortOrder: getDefaultSortOrder("name", sorter),
-      },
-      {
-        key: "maintenance_date",
-        title: t("hardware.label.field.maintenance_date"),
-        render: (value: IHardware) =>
-          value ? <DateField format="LL" value={value && value.date} /> : "",
-        defaultSortOrder: getDefaultSortOrder("maintenance_date", sorter),
-      },
-      {
-        key: "maintenance_cycle",
-        title: t("hardware.label.field.maintenance_cycle"),
-        render: (value: string) => <TextField value={value && value} />,
-        defaultSortOrder: getDefaultSortOrder("maintenance_cycle", sorter),
-      },
-      {
-        key: "image",
-        title: t("hardware.label.field.image"),
-        render: (value: string) => {
-          return value ? (
-            <Image width={80} alt="" height={"auto"} src={value} />
-          ) : (
-            ""
-          );
-        },
-      },
-      {
-        key: "asset_tag",
-        title: t("hardware.label.field.propertyCard"),
-        render: (value: string) => <TextField value={value ? value : 0} />,
-        defaultSortOrder: getDefaultSortOrder("asset_tag", sorter),
-      },
-      {
-        key: "serial",
-        title: t("hardware.label.field.serial"),
-        render: (value: string) => <TextField value={value ? value : 0} />,
-        defaultSortOrder: getDefaultSortOrder("serial", sorter),
-      },
-      {
-        key: "model",
-        title: t("hardware.label.field.propertyType"),
-        render: (value: IHardwareResponse) => (
-          <TagField value={value && value.name} />
-        ),
-        defaultSortOrder: getDefaultSortOrder("model.name", sorter),
-      },
-      {
-        key: "category",
-        title: t("hardware.label.field.category"),
-        render: (value: IHardwareResponse) => (
-          <TagField value={value && value.name} />
-        ),
-        defaultSortOrder: getDefaultSortOrder("category.name", sorter),
-        filters: filterCategory,
-        onFilter: (value: number, record: IHardwareResponse) => {
-          return record.category.id === value;
-        },
-      },
-      {
-        key: "status_label",
-        title: t("hardware.label.field.status"),
-        render: (_: string, record: IHardwareResponse) => {
-          const maintenanceTime = new Date(
-            record.maintenance_date?.date as string
-          );
-          const now = new Date();
-
-          const diffInMs = maintenanceTime.getTime() - now.getTime();
-          const diffInDays = Math.ceil(diffInMs / (1000 * 60 * 60 * 24));
-
-          let background = "";
-          let label = "";
-
-          if (diffInDays < 0) {
-            background = "red";
-            label = t("hardware.label.field.overdue");
-          } else if (diffInDays <= 10) {
-            background = "yellow";
-            label = t("hardware.label.field.pending");
-          }
-
-          return (
-            <TagField
-              value={label}
-              style={{
-                background,
-                color: background === "yellow" ? "black" : "white",
-              }}
-            />
-          );
-        },
-        defaultSortOrder: getDefaultSortOrder("status_label.name", sorter),
-        filters: filterStatus_Label,
-      },
-      {
-        key: "assigned_to",
-        title: t("hardware.label.field.checkoutTo"),
-        render: (value: IHardwareResponse) => (
-          <TextField strong value={value ? value.name : ""} />
-        ),
-        defaultSortOrder: getDefaultSortOrder("assigned_to.name", sorter),
-      },
-      {
-        key: "location",
-        title: t("hardware.label.field.rtd_location"),
-        render: (value: IHardwareResponse) => (
-          <TextField value={value && value.name} />
-        ),
-        defaultSortOrder: getDefaultSortOrder("location.name", sorter),
-      },
-      {
-        key: "rtd_location",
-        title: t("hardware.label.field.locationFix"),
-        render: (value: IHardwareResponse, record: IHardwareResponse) => (
-          <TextField
-            value={value && value.name}
-            onClick={() => {
-              list(
-                `location_details?id=${value.id}&name=${value.name}&status_id=${record.status_label.id}`
-              );
-            }}
-            style={{ cursor: "pointer", color: "rgb(36 118 165)" }}
-          />
-        ),
-        defaultSortOrder: getDefaultSortOrder("rtd_location.name", sorter),
-      },
-      {
-        key: "manufacturer",
-        title: t("hardware.label.field.manufacturer"),
-        render: (value: IHardwareResponse, record: IHardwareResponse) => (
-          <TextField
-            value={value && value.name}
-            onClick={() => {
-              list(
-                `manufactures_details?id=${value.id}&name=${value.name}&status_id=${record.status_label.id}`
-              );
-            }}
-            style={{ cursor: "pointer", color: "rgb(36 118 165)" }}
-          />
-        ),
-        defaultSortOrder: getDefaultSortOrder("manufacturer.name", sorter),
-      },
-      {
-        key: "supplier",
-        title: t("hardware.label.field.supplier"),
-        render: (value: IHardwareResponse, record: IHardwareResponse) => (
-          <div
-            dangerouslySetInnerHTML={{ __html: `${value ? value.name : ""}` }}
-            onClick={() => {
-              list(
-                `supplier_details?id=${value.id}&name=${value.name}&status_id=${record.status_label.id}`
-              );
-            }}
-            style={{ cursor: "pointer", color: "rgb(36 118 165)" }}
-          />
-        ),
-        defaultSortOrder: getDefaultSortOrder("supplier.name", sorter),
-      },
-      {
-        key: "purchase_date",
-        title: t("hardware.label.field.dateAdd"),
-        render: (value: IHardware) =>
-          value ? (
-            <DateField format="LL" value={value ? value.date : ""} />
-          ) : (
-            ""
-          ),
-        defaultSortOrder: getDefaultSortOrder("warranty_expires.date", sorter),
-      },
-      {
-        key: "order_number",
-        title: t("hardware.label.field.orderNumber"),
-        render: (value: string) => <TextField value={value ? value : ""} />,
-        defaultSortOrder: getDefaultSortOrder("order_number", sorter),
-      },
-      {
-        key: "warranty_months",
-        title: t("hardware.label.field.insurance"),
-        render: (value: string) => <TextField value={value ? value : ""} />,
-        defaultSortOrder: getDefaultSortOrder("warranty_months", sorter),
-      },
-      {
-        key: "warranty_expires",
-        title: t("hardware.label.field.warranty_expires"),
-        render: (value: IHardware) =>
-          value ? <DateField format="LLL" value={value && value.date} /> : "",
-      },
-      {
-        key: "notes",
-        title: t("hardware.label.field.note"),
-        render: (value: string) => (
-          <div dangerouslySetInnerHTML={{ __html: `${value ? value : ""}` }} />
-        ),
-        defaultSortOrder: getDefaultSortOrder("notes", sorter),
-      },
-      {
-        key: "checkout_counter",
-        title: t("hardware.label.field.checkout_counter"),
-        render: (value: number) => <TextField value={value ? value : 0} />,
-        defaultSortOrder: getDefaultSortOrder("checkout_counter", sorter),
-      },
-      {
-        key: "checkin_counter",
-        title: t("hardware.label.field.checkin_counter"),
-        render: (value: number) => <TextField value={value ? value : 0} />,
-        defaultSortOrder: getDefaultSortOrder("checkin_counter", sorter),
-      },
-      {
-        key: "requestable",
-        title: t("hardware.label.field.requestable"),
-        render: (value: string) => <TextField value={value ? value : 0} />,
-        defaultSortOrder: getDefaultSortOrder("requestable", sorter),
-      },
-      {
-        key: "assigned_status",
-        title: t("hardware.label.field.condition"),
-        render: (value: number) => (
-          <TagField
-            value={getAssetAssignedStatusDecription(value)}
-            style={{
-              background: getBGAssetAssignedStatusDecription(value),
-              color: "white",
-            }}
-          />
-        ),
-        defaultSortOrder: getDefaultSortOrder("assigned_status", sorter),
-        filters: filterAssignedStatus,
-        onFilter: (value: number, record: IHardwareResponse) =>
-          record.assigned_status === value,
-      },
-      {
-        key: "created_at",
-        title: t("hardware.label.field.dateCreate"),
-        render: (value: IHardware) =>
-          value ? (
-            <DateField format="LLL" value={value && value.datetime} />
-          ) : (
-            ""
-          ),
-        defaultSortOrder: getDefaultSortOrder("created_at.datetime", sorter),
-      },
-    ],
-    [filterCategory]
-  );
-
-  const handleCreate = () => {
-    handleOpenModel();
-  };
-
-  const handleOpenModel = () => {
-    setIsModalVisible(!isModalVisible);
-  };
-
-  const handleSearch = () => {
-    handleOpenSearchModel();
-  };
-
-  const handleOpenSearchModel = () => {
-    setIsSearchModalVisible(!isSearchModalVisible);
-  };
+  const collumns = useHardwareColumns({
+    sorter,
+    t,
+    list,
+    filterCategory,
+    filterStatus_Label,
+  });
 
   const refreshData = () => {
     tableQueryResult.refetch();
@@ -649,18 +228,6 @@ export const HardwareListMaintenance: React.FC<
     refreshData();
   }, [isEditModalVisible]);
 
-  useEffect(() => {
-    refreshData();
-  }, [isCloneModalVisible]);
-
-  useEffect(() => {
-    refreshData();
-  }, [isCheckoutModalVisible]);
-
-  useEffect(() => {
-    refreshData();
-  }, [isCheckinModalVisible]);
-
   const onCheckItem = (value: any) => {
     if (collumnSelected.includes(value.key)) {
       setColumnSelected(
@@ -678,33 +245,6 @@ export const HardwareListMaintenance: React.FC<
     );
   }, [collumnSelected]);
 
-  const listenForOutsideClicks = (
-    listening: boolean,
-    setListening: (arg0: boolean) => void,
-    menuRef: { current: any },
-    setIsActive: (arg0: boolean) => void
-  ) => {
-    if (listening) return;
-    if (!menuRef.current) return;
-    setListening(true);
-    [`click`, `touchstart`].forEach((type) => {
-      document.addEventListener(`click`, (event) => {
-        const current = menuRef.current;
-        const node = event.target;
-        if (current && current.contains(node)) return;
-        setIsActive(false);
-      });
-    });
-  };
-
-  useEffect(() => {
-    const aboutController = new AbortController();
-    listenForOutsideClicks(listening, setListening, menuRef, setIsActive);
-    return function cleanup() {
-      aboutController.abort();
-    };
-  }, []);
-
   // --- Loading and refresh logic ---
   const [loading, setLoading] = useState(false);
   const handleRefresh = () => {
@@ -717,7 +257,6 @@ export const HardwareListMaintenance: React.FC<
   };
 
   const pageTotal = tableProps.pagination && tableProps.pagination.total;
-  const { RangePicker } = DatePicker;
 
   const searchValuesByDateFrom = useMemo(() => {
     return localStorage.getItem("purchase_date_maintenance")?.substring(0, 10);
@@ -768,7 +307,6 @@ export const HardwareListMaintenance: React.FC<
       },
     ],
   });
-  const { Option } = Select;
 
   const handleChangeLocation = (value: number) => {
     if (value === 0) {
@@ -803,194 +341,51 @@ export const HardwareListMaintenance: React.FC<
       }}
     >
       <div className="search">
-        <Form
-          {...searchFormProps}
-          initialValues={{
-            location: localStorage.getItem("rtd_location_id_maintenance")
-              ? searchValuesLocation
-              : Number(rtd_location_id),
-            purchase_date:
-              localStorage.getItem("purchase_date_maintenance") !== null
-                ? searchValuesByDateFrom !== "" && searchValuesByDateTo !== ""
-                  ? [
-                      moment(searchValuesByDateFrom),
-                      moment(searchValuesByDateTo),
-                    ]
-                  : dateFromParam && dateToParam
-                    ? [moment(dateFromParam), moment(dateToParam)]
-                    : ""
-                : "",
-          }}
-          layout="vertical"
-          onValuesChange={() => searchFormProps.form?.submit()}
-          className="search-month-location"
-        >
-          <Form.Item
-            label={t("hardware.label.title.time")}
-            name="purchase_date"
-          >
-            <RangePicker
-              onChange={handleChangePickerByMonth}
-              format={dateFormat}
-              placeholder={[
-                `${t("hardware.label.field.start-date")}`,
-                `${t("hardware.label.field.end-date")}`,
-              ]}
-            />
-          </Form.Item>
-          <Form.Item
-            label={t("hardware.label.title.location")}
-            name="location"
-            className={
-              searchValuesLocation !== 0
-                ? "search-month-location-null"
-                : "search-month-location-null"
-            }
-          >
-            <Select onChange={handleChangeLocation} placeholder={t("all")}>
-              <Option value={0}>{t("all")}</Option>
-              {locationSelectProps.options?.map((item: any) => (
-                <Option value={item.value} key={item.value}>
-                  {item.label}
-                </Option>
-              ))}
-            </Select>
-          </Form.Item>
-        </Form>
+        <SearchFilterForm
+          searchFormProps={searchFormProps}
+          locationSelectProps={locationSelectProps}
+          handleChangePickerByMonth={handleChangePickerByMonth}
+          handleChangeLocation={handleChangeLocation}
+          searchValuesLocation={searchValuesLocation}
+          searchValuesByDateFrom={searchValuesByDateFrom}
+          searchValuesByDateTo={searchValuesByDateTo}
+          rtd_location_id={rtd_location_id}
+          dateFromParam={dateFromParam}
+          dateToParam={dateToParam}
+        />
+
         <div className="all">
           <TableAction searchFormProps={searchFormProps} />
-          <div className="other_function">
-            <div className="menu-container" ref={menuRef}>
-              <div>
-                <button
-                  className="menu-trigger"
-                  style={{
-                    borderTopLeftRadius: "3px",
-                    borderBottomLeftRadius: "3px",
-                  }}
-                >
-                  <Tooltip
-                    title={t("hardware.label.tooltip.refresh")}
-                    color={"#108ee9"}
-                  >
-                    <SyncOutlined
-                      onClick={handleRefresh}
-                      style={{ color: "black" }}
-                    />
-                  </Tooltip>
-                </button>
-              </div>
-              <div>
-                <button onClick={onClickDropDown} className="menu-trigger">
-                  <Tooltip
-                    title={t("hardware.label.tooltip.columns")}
-                    color={"#108ee9"}
-                  >
-                    <MenuOutlined style={{ color: "black" }} />
-                  </Tooltip>
-                </button>
-              </div>
-              <nav className={`menu ${isActive ? "active" : "inactive"}`}>
-                <div className="menu-dropdown">
-                  {collumns.map((item) => (
-                    <Checkbox
-                      className="checkbox"
-                      key={item.key}
-                      onChange={() => onCheckItem(item)}
-                      checked={collumnSelected.includes(item.key)}
-                    >
-                      {item.title}
-                    </Checkbox>
-                  ))}
-                </div>
-              </nav>
-            </div>
-            <div>
-              <button
-                className="menu-trigger"
-                style={{
-                  borderTopRightRadius: "3px",
-                  borderBottomRightRadius: "3px",
-                }}
-              >
-                <Tooltip
-                  title={t("hardware.label.tooltip.search")}
-                  color={"#108ee9"}
-                >
-                  <FileSearchOutlined
-                    onClick={() => setIsSearchModalVisible(true)}
-                    style={{ color: "black" }}
-                  />
-                </Tooltip>
-              </button>
-            </div>
-          </div>
+          <ToolbarActions
+            columns={collumns}
+            selectedColumns={collumnSelected}
+            onToggleColumn={onCheckItem}
+            onRefresh={handleRefresh}
+            onOpenSearch={() => setIsSearchModalVisible(true)}
+            t={t}
+          />
         </div>
       </div>
-      <MModal
-        title={t("hardware.label.title.search_advanced")}
-        setIsModalVisible={setIsSearchModalVisible}
-        isModalVisible={isSearchModalVisible}
-      >
-        <HardwareSearch
-          isModalVisible={isSearchModalVisible}
-          setIsModalVisible={setIsSearchModalVisible}
-          searchFormProps={searchFormProps}
-        />
-      </MModal>
-      <MModal
-        title={t("hardware.label.title.create")}
-        setIsModalVisible={setIsModalVisible}
+      <ModalsWrapper
+        t={t}
         isModalVisible={isModalVisible}
-      >
-        <HardwareCreate
-          setIsModalVisible={setIsModalVisible}
-          isModalVisible={isModalVisible}
-        />
-      </MModal>
-      <MModal
-        title={t("hardware.label.title.edit")}
-        setIsModalVisible={setIsEditModalVisible}
-        isModalVisible={isEditModalVisible}
-      >
-        <HardwareEdit
-          isModalVisible={isEditModalVisible}
-          setIsModalVisible={setIsEditModalVisible}
-          data={detail}
-        />
-      </MModal>
-      <MModal
-        title={t("hardware.label.title.checkout")}
-        setIsModalVisible={setIsCheckoutModalVisible}
-        isModalVisible={isCheckoutModalVisible}
-      >
-        <HardwareCheckout
-          isModalVisible={isCheckoutModalVisible}
-          setIsModalVisible={setIsCheckoutModalVisible}
-          data={detailCheckout}
-        />
-      </MModal>
-      <MModal
-        title={t("hardware.label.title.detail")}
-        setIsModalVisible={setIsShowModalVisible}
-        isModalVisible={isShowModalVisible}
-      >
-        <HardwareShow
-          setIsModalVisible={setIsShowModalVisible}
-          detail={detail}
-        />
-      </MModal>
-      <MModal
-        title={t("hardware.label.title.checkin")}
-        setIsModalVisible={setIsCheckinModalVisible}
-        isModalVisible={isCheckinModalVisible}
-      >
-        <HardwareCheckin
-          isModalVisible={isCheckinModalVisible}
-          setIsModalVisible={setIsCheckinModalVisible}
-          data={detailCheckin}
-        />
-      </MModal>
+        setIsModalVisible={setIsModalVisible}
+        isEditModalVisible={isEditModalVisible}
+        setIsEditModalVisible={setIsEditModalVisible}
+        isShowModalVisible={isShowModalVisible}
+        setIsShowModalVisible={setIsShowModalVisible}
+        isSearchModalVisible={isSearchModalVisible}
+        setIsSearchModalVisible={setIsSearchModalVisible}
+        isCheckoutModalVisible={isCheckoutModalVisible}
+        setIsCheckoutModalVisible={setIsCheckoutModalVisible}
+        isCheckinModalVisible={isCheckinModalVisible}
+        setIsCheckinModalVisible={setIsCheckinModalVisible}
+        detail={detail}
+        detailCheckin={detailCheckin}
+        detailCheckout={detailCheckout}
+        searchFormProps={searchFormProps}
+      />
+
       <TotalDetail
         filters={filters}
         links={HARDWARE_TOTAL_DETAIL_API}
@@ -1004,77 +399,17 @@ export const HardwareListMaintenance: React.FC<
           />
         </div>
       ) : (
-        <Table
-          className={(pageTotal as number) <= 10 ? "list-table" : ""}
-          {...tableProps}
-          rowKey="id"
-          scroll={{ x: 1850 }}
-          pagination={
-            (pageTotal as number) > 10
-              ? {
-                  position: ["topRight", "bottomRight"],
-                  total: pageTotal ? pageTotal : 0,
-                  showSizeChanger: true,
-                }
-              : false
-          }
-        >
-          {collumns
-            .filter((collumn) => collumnSelected.includes(collumn.key))
-            .map((col) => (
-              <Table.Column
-                dataIndex={col.key}
-                {...(col as any)}
-                key={col.key}
-                sorter
-              />
-            ))}
-          <Table.Column<IHardwareResponse>
-            title={t("table.actions")}
-            dataIndex="actions"
-            render={(_, record) => (
-              <Space>
-                <Tooltip
-                  title={t("hardware.label.tooltip.viewDetail")}
-                  color={"#108ee9"}
-                >
-                  <ShowButton
-                    hideText
-                    size="small"
-                    recordItemId={record.id}
-                    onClick={() => show(record)}
-                  />
-                </Tooltip>
-
-                <Tooltip
-                  title={t("hardware.label.tooltip.edit")}
-                  color={"#108ee9"}
-                >
-                  <EditButton
-                    hideText
-                    size="small"
-                    recordItemId={record.id}
-                    onClick={() => edit(record)}
-                  />
-                </Tooltip>
-                <Tooltip
-                  title={t("hardware.label.tooltip.delete")}
-                  color={"red"}
-                >
-                  <DeleteButton
-                    resourceName={HARDWARE_API}
-                    hideText
-                    size="small"
-                    recordItemId={record.id}
-                    onSuccess={() => {
-                      setIsTotalDetailReload(!isTotalDetailReload);
-                    }}
-                  />
-                </Tooltip>
-              </Space>
-            )}
-          />
-        </Table>
+        <HardwareTable
+          columns={collumns}
+          selectedColumns={collumnSelected}
+          tableProps={tableProps}
+          onShow={show}
+          onEdit={edit}
+          onDeleteSuccess={() => setIsTotalDetailReload(!isTotalDetailReload)}
+          pageTotal={pageTotal}
+          t={t}
+          resourceName={HARDWARE_API}
+        />
       )}
     </List>
   );
